Moved summarize files to appropriate directories
