--- summarizer/expected.txt
@@ -1,1 +1,0 @@
-Bad news first because I’m a cup half empty kind of guy: I gave up on working with PHP and Laurel because I just ran out of time and lack the drive to really focus on learning a new technology when there are so many other fun things waiting to be tinkered.The latter was no fun so I’ll just talk about the react component.One was a new component for a react-app while the other was just a simple “add a button to html page” (Sorry, I ran out of time).So I just those code segments into the project and submitted the PR.This took some time as I was no longer just copy, pasting and editing.
